Organize react-ui primitives by domain
