+++ packages/react-ui/src/chat/index.ts
@@ -1,0 +1,9 @@
+export type {
+  ChatController,
+  ChatProviderProps,
+  ComposerContextValue,
+  ThreadContextValue,
+} from "../contexts";
+export { ChatProvider, useChatContext, useComposer, useThread } from "../contexts";
+export { Composer } from "./composer";
+export { Thread } from "./thread";
