+++ packages/react-ui/src/primitives/index.ts
@@ -1,0 +1,3 @@
+export { composeRefs } from "./refs";
+export type { PrimitiveProps, PrimitiveRef } from "./render";
+export { renderPrimitive } from "./render";
