+++ packages/react-ui/test/exports.test.ts
@@ -1,0 +1,30 @@
+import { describe, expect, expectTypeOf, it } from "vitest";
+
+import { ChatProvider, Composer, Thread } from "../src/chat";
+import { Completion, CompletionProvider } from "../src/completion";
+import { HumanInput } from "../src/human-input";
+import type { MessageToolPart } from "../src/message";
+import { Message } from "../src/message";
+import {
+  ChatProvider as SharedChatProvider,
+  CompletionProvider as SharedCompletionProvider,
+} from "../src/shared";
+
+describe("public entrypoints", () => {
+  it("exports namespace barrels from subpaths", () => {
+    expect(Thread.Root).toBeTypeOf("object");
+    expect(Composer.Root).toBeTypeOf("object");
+    expect(Message.Root).toBeTypeOf("object");
+    expect(HumanInput.Approvals).toBeTypeOf("object");
+    expect(Completion.Root).toBeTypeOf("object");
+  });
+
+  it("keeps shared provider exports aligned with domain barrels", () => {
+    expect(SharedChatProvider).toBe(ChatProvider);
+    expect(SharedCompletionProvider).toBe(CompletionProvider);
+  });
+
+  it("exports public helper types from domain barrels", () => {
+    expectTypeOf<MessageToolPart>().toMatchTypeOf<{ type: "tool" }>();
+  });
+});
